Define %protoOfField in the interpreter directly
Its definition in s5 requires a recursive function, which will
inevitably lead to precision problems. With the rec-768e726 test and
m-CFA with m = 1, this is already the case. Defining it directly in the
abstract interpreter itself seems like the only solution. However,
there are 65 more recursive functions defined in S5's environment file,
and this could mean that those 65 functions should also be redefined
inside the abstract interpreter.